refactor: replace badge by status indicator
